Equity folder, about to muck with scenarios
