+++ ui/src/app.jsx
@@ -11,7 +11,7 @@ import VirtualSchoolPage from './virtual_school/virtual_school_page.jsx';
 import HomePage from './home/home_page.jsx';
 import DemosPage from './home/demos_page.jsx';
 import CsBiasPage from './home/cs_bias_page.jsx';
-import EquityPage from './home/equity_page.jsx';
+import FairPage from './message_popup/equity/fair_page.jsx';
 import ConsentPage from './home/consent_page.jsx';
 import * as MessagePopup from './message_popup/index.js';
 
@@ -42,7 +42,8 @@ export default React.createClass({
     '/bias': 'biasHome',
 
     // For CSS workshop
-    '/equity': 'equityPage',
+    '/equity': 'fairPage',
+    '/equity/fair': 'fairPage',
     
     // Included in /demos, shared externally, or used in playtests
     '/teachermoments/original': 'messagePopup',
@@ -226,8 +227,8 @@ export default React.createClass({
     return <MessagePopup.HMTCAExperiencePage query={query} />;
   },
 
-  equityPage(query = {}) {
-    return <EquityPage query={query} />;
+  fairPage(query = {}) {
+    return <FairPage query={query} />;
   },
 
   messagePopupPlaytest(cohortKey, query = {}) {
